+++ src/variables.js
@@ -1,7 +1,5 @@
-var dhbanner = null;
 var dhvtcbg = null;
 var dhcustombg = (localStorage.getItem("custom-background") !== null ? localStorage.getItem("custom-background") : "");
-exports.dhbanner = dhbanner;
 exports.dhvtcbg = dhvtcbg;
 exports.dhcustombg = dhcustombg;